Forward Intercom component props to intercomSettings and handle mobile
All props provided to the Intercom component are deemed to be the
source of truth for styling those aspects of the widget. This
ensures there's no style mismatch between the facade and the
loaded widget. A styling mismatch can still occur if a custom
icon is set in Intercom.

To mirror the default Intercom widget behavior padding props
are only applied on screen sizes with a width greater than 900px.
For smaller screen sizes the default padding of 20px is used.

Fix incorrect alignment proptypes default prop value